Added "mousekeys" from this patch:
https://github.com/LukeSmithxyz/st/blob/patched/patches/st-scrollback-mouse-20161020-6e79e83.diff

This lets you bind a function to a mouse button (or scrollwheel) - I use
it for scrollback functionality.

This commit makes gcc give a couple of warnings, which can be ignored.
The warnings happen because of the way the program is structured. The
simplest way I can see for fixing it would be to move the functions
kscrollup and kscrolldown to the x.c file, but I would like to avoid
that since x.c seems to be strictly for X related code, and those
functions would thus be misplaced in that file. No big deal any which
way, but I'd still like to get rid of the warnings.